@@ -3,32 +3,30 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.13.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/js-cookie": "^3.0.1",
-    "@types/node": "^12.20.15",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.7",
-    "axios": "^0.24.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/jest": "^29.5.14",
+    "@types/js-cookie": "^3.0.6",
+    "@types/node": "^22.9.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "axios": "^1.7.7",
     "http-proxy-middleware": "^0.19.1",
-    "js-cookie": "^3.0.1",
-    "jwt-decode": "^3.1.2",
+    "js-cookie": "^3.0.5",
+    "jwt-decode": "^4.0.0",
     "normalize.css": "^8.0.1",
     "plyr": "3.6.8",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-cookie": "^4.0.3",
-    "react-device-detect": "^1.17.0",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.12.0",
-    "react-progressive-graceful-image": "^0.6.14",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.3.1",
+    "react-cookie": "^7.2.2",
+    "react-device-detect": "^2.2.3",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.3.0",
+    "react-progressive-graceful-image": "^0.7.0",
+    "react-router-dom": "^6.28.0",
     "react-scripts": "5.0.1",
-    "typescript": "^4.3.2",
-    "typescript-json-serializer": "^3.4.0",
-    "use-resize-observer": "^8.0.0"
+    "typescript": "^4.9.5",
+    "typescript-json-serializer": "^6.0.1",
+    "use-resize-observer": "^9.1.0"
   },
   "scripts": {
     "start": "BROWSER=none react-scripts start",
