@@ -1,4 +1,4 @@
-import { BrowserRouter, Route, Switch, useParams } from 'react-router-dom';
+import {BrowserRouter, Route, Routes, useParams} from 'react-router-dom';
 import Editor from './pages/Editor';
 import Compilation from './pages/Compilation';
 import Main from './pages/Main';
@@ -9,13 +9,13 @@ function App() {
   return (
     <div className="app-root">
       <BrowserRouter>
-          <Switch>
-            <Route exact path="/" component = { Main }  />
-            <Route path="/search"  component = { Main } />
-            <Route path="/editor/:id" children = { <EditorRouter /> } />
-            <Route exact path="/video-wall" component = { VideoWall } />
-            <Route exact path="/compilation" children ={ <Compilation /> } />
-          </Switch>
+          <Routes>
+            <Route path="/" element = { <Main /> }  />
+            <Route path="/search"  element = { <Main /> } />
+            <Route path="/editor/:id" element = { <EditorRouter /> } />
+            <Route path="/video-wall" element = { <VideoWall /> } />
+            <Route path="/compilation" element = { <Compilation /> } />
+          </Routes>
       </BrowserRouter>
 
     </div>
@@ -25,7 +25,9 @@ function App() {
 function EditorRouter() {
   let { id } = useParams<{ id: string }>();
   return (
-    <Editor videoId={id} />
+      <>
+          { id && <Editor videoId = { id } /> }
+      </>
   );
 }
 
